// src/enums/messages.ts
export enum Messages {
  USER_CREATED = "USER_CREATED",
  USER_EXIST = "USER_EXIST",
  UNKNOWN_USER = "UNKNOWN_USER",
  EMAIL_NOT_EXIST = "EMAIL_NOT_EXIST",
  INVALID_PASSWORD = "INVALID_PASSWORD",
  BUSINESS_CREATED = "BUSINESS_CREATED",
  UNKNOWN_ERROR = "UNKNOWN_ERROR",
  LOGIN_SUCCESSFULLY = "LOGIN_SUCCESSFULLY",
  LOGIN_FAILED = "LOGIN_FAILED",
  PRODUCT_CREATED = "PRODUCT_CREATED",
  UNAUTHORIZED = "UNAUTHORIZED",
  INVALID_TOKEN = "INVALID_TOKEN",
  AUTHORIZED = "AUTHORIZED",
  EXPIRED_TOKEN = "EXPIRED_TOKEN",
}